refactor(viewer): split receipt projection support
